requirements.txt
streamlit
pandas
supabase
google-genai
altair
python-dotenv
pypdf
python-docx
